+++ src/Symfony/Bundle/SecurityBundle/CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
  * Make `Security::login()` return the authenticator response
  * Deprecate the `security.firewalls.logout.csrf_token_generator` config option, use `security.firewalls.logout.csrf_token_manager` instead
  * Make firewalls event dispatcher traceable on debug mode
+ * Add `TokenHandlerFactoryInterface`, `OidcUserInfoTokenHandlerFactory`, `OidcTokenHandlerFactory` and `ServiceTokenHandlerFactory` for `AccessTokenFactory`
 
 6.2
 ---